added structure
